Create signup.fxml
This is the fxml file for the signup scene, which is used when the user hits the sign up button in the main scene. The user is asked to enter an email and a username that haven't been already used, to choose a password and confirm it, and to confirm they are not a bot through a primitive CAPTCHA. If successful, the information is stored in a mySQL database, and the user is directed back to the first scene to login with their new account.