Create Server.java
This server class implements the notion of multi-threading. Now, the application works using threads. Each client has a thread to communicate with this server. The former issues a request, which the server translates, processes in accordance with the SQL database, and sends back a response. So if I want to login, I plug in my credentials, the server verifies them, and gives me permission to access the second scene. Etc...